ci: test the manifest generation
The script osbuild-image-test can dynamically filter based on the
machine characteristics. We use an intermediary program to extracts the
distribution name from /etc/os-release and the arch from os.platform.
This way we can run the CI on each kind of machines/distros.

The CI is only used to generate new image-infos that can later be
integrated to the database.