tools/osbuild-image-test: fail upon error
The program now returns a code greater than zero if an error occurred
during the processing of the input files.

Is considered an error and image-info difference between the source test
case and the result after osbuild ran the image.

Is considered a warning a missing image-info from the test case and a
manifest validation error from the test case.