gh: automatically include the new image-info
When a PR will land on the manifest-db project, after Schutzbot is
successful, this action will start. It will pull the latest compiled
image-info and create a new commit on top of the PR with them. If there
are changes, the PR will then start a new CI cycle, this time, it will
have nothing to commit.