auto-update: only import CI's distros/archs
If the bot would import manifests that would not be computed on the CI,
we would run into an issue: Getting the manifest side of the db out of
sync with image-info side for the ditros-archs we can't run.
Instead, make the tooling only import the manifests it is able to handle.
We get the list of computable distros and arch from the gitlab yml file,
this way, nothing has to change in the tooling when we update the list
of distros in the future.