schutzbot: clean SELinux mismatch
There's a difference in SELinux mismatch between OSBuild'CI and
manifest-db'CI. Two issues are involved in this issue:

* I have missing packages in my setup and that leads to SELinux mismatch
  between what I'm computing on this CI and OSBuild's CI.

* The which command used alongside matchpathcon does not give the same
  result when its called from root or user context. And therefore we
don't have access to the proper label for the image-info command.

The first issue is fixed by by copying the deploy.sh setup procedure
from OSBuild.

The second issue is fixed by specifying the path to access
the osbuild executable when calling matchpathcon.