auto_update: generate the manifests
Introduces manifest generation as a preliminary step during Gitlab CI
before building the image-info. This removes the need of generating the
manifests in osbuild-composer.

Also this simplifies the workflow to generate an update by
only keeping in the trigger in the Github action (before we were having
copying happening there).