auto_update: no branch creation in github action
Remove the branch creation from the github action and trigger the update
on the main branch. The branch is created in the last step of the
manifest generation pipeline.

If a user wants to test the pipeline on another set of commits, they'll
have to push to gitlab on a branch and then trigger the pipeline on this
branch.

You can use this command to trigger the pipeline on a branch $BRANCH
curl --request POST --form token=$TRIGGER_TOKEN --form ref=$BRANCH https://gitlab.com/api/v4/projects/36844106/trigger/pipeline