image_info/commons: framework to load the Commons
Most of the data that goes in the image-info comes from Common elements
to all the image types.

This commit brings up a mechanisms to dynamically load all the modules
under the common folder. Then, when needed, the Target iterates on every
class that inherit from Common and initializes them with the needed
data, either the json, or the tree.