image_info/report: define how to compose a report
Contains all the definition of report elements that can be composed in a
report. The vast majority of the classes (at the exception of Report and
ReportElement) will inherit from ReportElement and thus can be loaded
from a JSON File.

ReportElements are using the attr libraries. Elements can be transformed
to dict easily with the attr.asdict function.