report: Common be InsideTreeElement
There is two kinds of elements to be added to a report. The elements who
needs to have a Tree to explore and the others. This commit replace the
previous notion of "Common" which was to define the report pieces
included in all kind of images all the time by the notion of "Inside a
tree" and reorganize the code base accordingly.

A follow up will be to extract in another folder the things that must be
done outside the tree.