[WIP]core: splitting mounting and exporting
Before we were having specialized ReportElement that were responsible
for both discovering the values from the partition table in order to
export it in the report and also to mount the File System for the rest
of the report elements to be able to discover what they needed to.

And this was bad because we were mixing two different kind of
responsibilities. In a future version, we might want to rely on OSBuild
to perform the actual mounting of the FS instead of maintaining custom
code. Especially if we need to support other kinds of FS (e.g btrfs).