db: store the filename in the database.
Before the image filename was built on the fly by the testing pipeline.
This was working under the assumption that if no image name was found in
the export stage, "image.raw" will be the correct fallback. This was
functioning until the introduction of the `minimal_raw-boot` image
definitions that are saving their result on the filename `raw.img`.

To take these into account, upon importing the entries for the
database, include also the filename that's coming from the composer
request.

We also keep the previous method as a fallback for the DB entries that
will never get updated (e.g because of EOL distribution) in case anyone
wants to build them in the future.