tools/filterdiff: refactor the abstraction layer
Before the `load` and `diff` functions where used to load a DB entry or
a raw image-info and diff it. The code flow was not the best however.
Another issue was that using the filterdiff in other programs
(osbuild-image-test and update_tool) was not straightforward. For
instance, the logic to load the unwanted tmpfiles_d had to be
reimplemented each time.

Long story short, this commit refactors that. It brings up a new
ImageInfo class in the filterdiff library. An ImageInfo object can be
instanciated from a Dict containing the Image-info, but also a helper
class method enables loading from files.

Diffing two ImageInfo objetcs is done by invoking diff on one of the two
objects with the other one as parameters. This greatly simplify the
workflow when a diff is needed. Espcially when it comes to tmpfiles_d
merging.