schutzbot/deploy.sh: install only osbuild packages
None of the scripts in this repository actually run or use
osbuild-composer binaries of tools. Manifests are generated from the
osbuild/images repository and they are built directly using osbuild.

The only situation when osbuild-composer is somewhat used is during the
manifest generation. Specifically, we need to install all build
dependencies of osbuild-composer before the manifest generation for the
tools in osbuild/images to compile. The reason is that osbuild/images
does not come with a SPEC file to enable us not using osbuild-composer
at all.

Modify the `deploy.sh` script to install only osbuild packages.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>