manifest-db: make all UUIDs unique
We get regular failures in the manifest-db run. With the extra
visibility from osbuild#123
it seems the errors we get (sometimes) are related to an xfs mount
```
mount/- (org.osbuild.xfs): mounting /dev/98302bc3-1aa3-4eab-96e9-9f1c6eb539ce/rootlv -> /var/lib/osbuild/store/tmp/buildroot-tmp-3y2wyhos/mounts/
mount/- (org.osbuild.xfs): already unmounted: /var/lib/osbuild/store/tmp/buildroot-tmp-3y2wyhos/mounts/
```
which looks similar the UUID clashes we observed in
osbuild/osbuild#1641

Upon inspecting the manifests it seems we have quite a few UUIDs in
our manifests. Those *might* explain the issues that we sometimes
see because the UUIDs is global to the host so running multiple tests
in parallel can lead to clashes when xfs/btrfs assume that the
mount is already done when in fact it was a different loopdevice
that just happend to clash with the UUID.

This commit creates a new UUID for each file where a duplicated
UUID is found (but keeps that UUID inside the same file stable).

We could make this smaller by limiting this to only xfs/btfs which
(AFAIK) are the only systems affected by UUID clashes but my
script is not smart enough for this.